Entity placement introduced. Fighter makes damage when it crashes.
Placement is just an entity classification as ground/air/water. Currently it is only  used to replace the fighter crash condition. It was implemented by waiting for one second after destroyed, but it failed with tall buildings (fighters stucked into the building for one second before exploding). Now the fighter check not to crash with air entities (the desired behaviour).

The fighter does a damage equal to its max health when it crashes.