Speed param in route points. Route point panel in scenario editor.
Introduced new speed change param in the routes, it indicates the speed of the entity when in reaches that point.  It is especified as a factor of the entity max velocity, so a 1 is the entity normal velocity. 

The other route point param "pause" was incorrectly implemented, it was applying the pause of the next point instead of the current one. This has been fixed, the pauses and the speed changes are configured in the point they ocurr. The route pauses in the  level 1 boss route have been adapted to this change.

Speed param is currently only used by the ground boss type.