First shot delay changed. Turret min distance to target. Dreadnought tower holds fire.
First Shot delay now starts counting from the first visible time rather that from creation time to ensure all entities behave the same way. This is very important for planes, because depending on the formation there can be a significant differente in the interval from their creation to when they become visible.

Dreadnought tower is no lo longer based on staticstructure because it needs to hold fire until its parent has started its route.

Turret now implements a minimun distance to target (measured as the distance from the target to the intersection of the target's plane with the line from the camera to the turret)