Added difficulty level support. Added player profile. Fixed some bugs. Adjusted some entity parameters.
Difficulty level support:
	Difficulty levels are composed of two factors:
	- Player selected difficulty.
	- Scenario configured difficulty.

	This two factors are combined using configurable weights  into the difficulty factor using the following formula:
	base_difficulty+ (player_difficulty*player_weight)+(scenario_difficulty*scenario_weight).

	Current weight configuration is:
	 - Base difficulty: 0.8  
	 - Player factor: 0.075
	 - Scenario factor: 0.05

	The difficulty factor affects the following enemy parameters:
	- First shot delay (divided)
	- Time between shots (divided)
	- Enemy bullet speed (multiplied)
	- Health of all Bombers and Bosses (multiplied)

Player Profiles
	Player profiles are placeholders for the player configuration and data, they will hold:
		- Player name (for future rankings)
		- Keyboard mapping.
		- Player difficulty level.
	PlayerManager uses a profile to compute the current difficulty.

SetTarget specializations removed because now EntityBase's implementation suscribes to the target events and clears the target if the target is killed or removed.
Difficulty level integration in the game and in the scenario editor.
Fighter generated a core when it targeted an entity was marked to be removed. Fixed by chekcing the target's health. It did only happened in the editors.
Some entity weapons have been copied from EnemyBasicWeapon to be able to fine tune the launching point (should not be necessary but currently is the only way to do it)
Scenario Editor improvements: Added GodMode and UseAmmo (for bombs) options. Volume and Mouse trace distance moved to the Options menu.