Automatic shadow preactivation.  Play area elements now support Start/Stop, replacing Reset. Negative delay support. Removed player resource preparation in PlayAreaManager.
Fixed the classic problem with shadows appearing suddenly.

Entities are activated before its activation time by estimating the length of the shadow  visible before the entity itself.

Shadow preactivation is simply the estimated shadow lenght in the X axis. 
Shadow lenght is computed by intersecting the XZ plane at the entity min Y with a line from the sun passing throught the entity highest height, that gives us the ammount of X we have to show the entity in advance.
It can only be negative or zero, shadows lengths>0 are clamped to 0 as it means that the shadow falls below the entity start (azimuth 90 - 270).
This computation takes into account both elevation and azimuth.
		 
This computation fails when the entity shadow is projected below its min Y but works most of the times. This problem can be fixed by tracing against the world, but it also creates problems in some other cases. So, to keep things as simple as possible the cases not covered by this computation are required to be fine tuned using a negative delay.
		
The implementation is based on GetDesignAABBox, which replaces GetDesignBBox, this new version takes into account the rotation of the models and child entities. This API gives an AxisAlignedBBox, so care must be taken when using it. It fits well to shadow prediction because the camera always move along the X axis, but this API should not be used in collision detection.

To able to precompute the shadow preactivation, playarealements now support Start/Stop, this two new APIs replace the old Reset.