Added support for variouos shading models, Gouraud, Phong and Balanced. Shading model can be changed in the scenario editor by pressing the 'M' key. Added suppport to reload the shaders pressing F12 in the Scenario Editor. Shaders now dump the source code when a compilation error is detected.
Balanced shading model is a compromise between Phong and Gouraud, it uses Phong for the light sun and Gouraud for dynamic lights. This allows the future development of normal maps and improves the fisual quality without affecting the performance too much.

Water is only be affected by lighting when phong model is used, in all the other models dynamic lights are processed in pixel shader.