Button sound support. Navigation through controls with keys. Dialogs restore the focus to the previously focused window.
Navigation is done amongst the focusable controls (that which return true in OnWantFocus). 
The user can navigate using the cursors and TAB/Shift+Tab.

Buttons can be configured to want focus to be able to navigate through them. 
When a button has the focus it can be pressed using SPACE and ENTER keys.
Buttons can be configured to take the focus on hover.
Buttons now implement hover and click sounds.