Configured sounds fx for all entities. Sounds channels are recicled. Sounds now honor the configured volume. Most sound fx restricted to 1 channel.
Sound recicling means that when all the sound channels are busy, the oldest sound channel is restarted. This is done by recovering the channel from the original sound and assigning it to the new one. This behaves far better that having sounds not playing at all because old sounds are already playing (a lot of times those old sounds cannot be heard because they are fading out).

Having only one channel per sound fx is far better than having a lot because otherwise explosions hide the music and other effects.