Fixed linux problems with  KDE and Xinerama/TwinView.
Because the Render Context has no relation with an XWindow and to avoid integration problems with window managers the X window is recreated each time it is toggled between window and fullscreen modes. This is by far the simplest solution, and works fine at least in gnome and kde.

Now the engine goes fullscreen to the first xinerama monitor to avoid getting the window splitted between monitors.

Now the engine window can directly start in fullscreen instead of always starting windowed and then toggled fullscreen.

Windows implementation adaptation is pending.