WIP: In progress Android port and shader only rendering.
Removed support for old technologies:
 - Fixed pipeline, the new implementation uses only shaders.
 - Vertex arrays. Now VBO are used using vertex attributes. Non VBO buffers are only used where required.
 - Render to texture using PBuffers and  the back buffer. Now only FBO is supported.
 - System font support, they are not portable. Now only texture fonts are used.

Port to Android 2.3.3 and above in progress. Ported as a NativeActivity using GLES-2.0 (shaders only).
 - When compiled for android the engine becomes monolithic due to a problem using RTTI across dynamic library boundaries. I do not know how to fix it yet.
- OpenAL is being used as an static library (some tune-up required).

TO-DO list is too big to put it here, enough to say that there are some things to be implemented and that the code includes dirty (and temporal) workarounds for several problems arised during the port.

This is just a safety commit :), too many changes to lose them.