WIP: Android port. GLES 1 Rendering. Basic input management.
Fixed pipeline rendering recovered to support GLES 1 devices. 
For the time being, fixed pipeline will be enabled with conditional compilation and it won't be used in desktop versions. It may be available in android version to create a unified GLES 1 / GLES 2 version, but this point has to be evaluated.

Fonts scaled to powers of 2. Pending to scale other textures.

The game now works in the emulator using GLES 1. 
  No shadow support in GLES 1.
  Key based input, virtual joystick not yet implemented.
  Mouse emulated using the touchpad.