Texture buffers are freed after uploading them to opengl.
By default, texture buffers are inmediatelly freed to reduce memory consumption.
In the few places where texture memory access is needed, the texture client explicitly loads the texture as resident, instead of deserializing the texture object.

The only places where resident textures are needed are:

 - Font: For character crop. Can be freed after cropping.
 - Scenario ColorMap. Must be kept.

Accesing opengl memory through the api is not an option because texture memory access is not implemented in GLES.

Scenario files updated to erase the serialized ColorMap texture.