Optimizations: GUI drawing. No model custom clipping if auto shadow volume is inactive.
Windows are now drawn on screen space, not on window space.
This causes windows to be a bit more difficult to draw but it allows us to get rid of all the setviewport, camera, and clipping sequence.
There is no window clipping, this can cause some visual gliches in the editors. If this becomes a problem clipping can be implemented as a configuration parameter in the GUIManager.

IMPORTANT: In order to restore the GUI projection and viewport, any window changing either of them must call RestoreViewport method of the GUIManager before exiting the draw call.

Custom model clipping deactivated if automatic shadow volume computation is not active.
PlayAreaManager and the entities themselves perform enough life time control to make this a waste of time when playing.