Deferred Render with three modes of operation. Underwater terrain performance improvement.
The deferred render now has three modes of operation, this makes the render complex, but makes it render in the fastest way depending on the rendered scene, the Water/NoWater separation is made to light the terrain and models below the water line (two layers of lighting due to transparent surfaces make deferred rendering quite complex):

The three modes of operation are:
- Forward: (same as ForwardShader) Used when there is no lighting, i.e. GUI elements.
- NoWater: Full deferred shader, used when there is no water in the scene. In this mode, the scene data is gathered into GBuffers and used with simple lighting shaders, one for the sun and another for the dynamic lights.
- Water: Half forward and half deferred, used when there is water in the scene. In this mode, the scene is rendered at the same time (single pass) to the GBuffers and to a buffer lighted y the sun. After this, the lighted buffer is copied into the screen frame buffer. The final step just adds the dynamic lights contribution from the data in the GBuffers.

Underwater terrain below the min fog is replaced with a clear operation when playing.
It is kept separated from the regular terrain sectors to be able to skip it while playing.
It has not been completely removed to show it in the scenario editor.