Depth buffer is now cleared explicitly. Clip rect is no longer implicitly restricted to the configured viewport. GameInterface was drawing nothing in some control paths.
Depth buffer is no longer cleared in the OpenGLViewport when the frame starts, it is now done on-demand. GUI does not use z buffering, so it does not need to clear the depth buffer, it is only really needed when drawing a non-gui scene.
A new DepthClear method has been added to do so, it is specially important to render multiple non-gui scenes at the same time, because otherwise the z values of both scenes can be mixed and clearing the depth buffer for every scene is extremely inefficient.

As a consequence clearing depth buffer and stablishing the clip rect is now responsability of the windows.

Clipping management has been adjusted and ensured it does not affects the shadow buffers.