Removed deprecated texture features. Delayed opengl setup of textures. Model textures are no longer registered by name.
AlphaFile and ColorKey support removed from the textures, both of them were superseded by the native alpha channel support of the png files.

Textures now can perform a delayed opengl setup to avoid problems when textures are loaded before a render context has been created. If the setup fails at load time it is retried the first time the texture is prepared for rendering.

Textures loades from models are no longer registered by name to avoid storing them in the GameResources.conf file when the EntityEditor saves it, this can lead to a few duplicated textures but is better than having to constantly monitoring the file. This is a side effect of having the entityeditor saving all named objects in the GameResources system, a cleaner solution has to be implemented, but for the time being this one is good enought.