Video user settings moved from the GUIManager to the game main window. GUIManager initialization sequence refactored to be able to specify video settings before the window is created.
UserSettingsFile/Node keys have been removed from the GUIManager. That implementation was temporal and has now been fixed. It was necesary to have the window created using the user settings because there was not possible to specify the ScreenProperties before the window was created.

To be able to remove these settings from GUIManager and handle them from the game main window, the initialization sequence has been modified introducing an aditional setup step.

Window auto registration mechanism has been removed and replaced with a named binding to the MainWindow object in the GameGUI system.

Now the initialization sequence is as follows:

- GameGUI Objects are instantiated and registered by name.
- GameGUI Objects are unserialized and object wrappers are dinamically bound to their target objects. 
  When unserialized, the main window configures: 
    GameGUIManager: Windowed/FullScreen mode, resolution, position and VSync mode.
    Render: Selected render path and shadow quality.
    SoundManager: Volume settings.
- GameGUIManager's Setup method is called. It executes the following steps:
    Creates the Viewport.
    Sets up the Render.
    Bounds to the main window by name (MainWindow) and initializes it.
- GameGUIManager's EnterGUILoop method is called, starting normal program execution.